fix: dataset import and architecture adjustments
This commit improves the import and data classification, training,
and validation from the dataset. The input shape from the image was
altered to accommodate the dataset at the "SkinDiseaseClassifier"
function. Also, all the paths to the dataset were correct in the
code, passing the path to the model.

Now, the code is robust and capable of training and validating the
skin classifier.